# openapi.yaml
components:
  schemas:
...
    DataResponseAuthData:
      type: object
      properties:
        data: {$ref: '#/components/schemas/AuthData'}
        error: {type: string}
        timestamp: {format: date-time, type: string}
...
    ListResponsePlaceDto:
      type: object
      properties:
        total: {format: int32, type: integer}
        perPage: {format: int32, type: integer}
        offset: {format: int32, type: integer}
        data:
          type: array
          items: {$ref: '#/components/schemas/PlaceDto'}
        error: {type: string}
        timestamp: {format: date-time, type: string}
...
    TagDto:
      description: Список городов
      type: object
      properties:
        id: {format: int32, description: Идентификатор тэга, type: integer}
        tag: {description: Собственно тэг, type: string}
    ListResponseAuthData:
      type: object
      properties:
        total: {format: int32, type: integer}
        perPage: {format: int32, type: integer}
        offset: {format: int32, type: integer}
        data:
          type: array
          items: {$ref: '#/components/schemas/AuthData'}
        error: {type: string}
        timestamp: {format: date-time, type: string}
...
    PlaceDto:
      description: Список городов
      type: object
      properties:
        id: {format: int32, description: Идентификатор города/страны, type: integer}
        title: {description: Название города/страны, type: string}
...
    AuthData:
      type: object
      properties:
        country: {type: string}
        city: {type: string}
        birth_date: {format: date-time, type: string}
        about: {type: string}
        photo: {type: string}
        last_name: {type: string}
        token: {type: string}
        reg_date: {format: date-time, type: string}
        is_blocked: {type: boolean}
        deleted: {type: boolean}
        messages_permission:
          type: string
          enum: [ALL, FRIENDS, NOBODY]
        last_online_time: {format: date-time, type: string}
        phone: {type: string}
        id: {format: int32, type: integer}
        first_name: {type: string}
        email: {type: string}
...
    DataResponseTagDto:
      type: object
      properties:
        data: {$ref: '#/components/schemas/TagDto'}
        error: {type: string}
        timestamp: {format: date-time, type: string}
    AccountResponse:
      type: object
      properties:
        data:
          additionalProperties: {type: string}
          type: object
        error: {type: string}
        timestamp: {format: date-time, type: string}
    ListResponseTagDto:
      type: object
      properties:
        total: {format: int32, type: integer}
        perPage: {format: int32, type: integer}
        offset: {format: int32, type: integer}
        data:
          type: array
          items: {$ref: '#/components/schemas/TagDto'}
        error: {type: string}
        timestamp: {format: date-time, type: string}
  securitySchemes:
    jwt: {scheme: bearer, in: header, name: Authorization, type: http}
servers:
- {description: Generated server url, url: 'http://localhost'}
openapi: 3.0.1
paths:
...
  /api/v1/users/me:
    get:
      summary: Получение текущего пользователя
      security:
      - jwt: []
      operationId: getMe
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/DataResponseAuthData'}
      tags: [Контроллер для работы с профилем]
    delete:
      summary: Удалить профиль пользователя
      security:
      - jwt: []
      operationId: deleteUser
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/AccountResponse'}
      tags: [Контроллер для работы с профилем]
    put:
      summary: Обновить профиль пользователя
      security:
      - jwt: []
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthData'}
        required: true
      operationId: updateUser
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/DataResponseAuthData'}
      tags: [Контроллер для работы с профилем]
  /api/v1/platform/cities:
    get:
      summary: Запрос списка городов в количестве itemPerPage содержащих city
      security:
      - jwt: []
      operationId: getCities
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/ListResponsePlaceDto'}
      parameters:
      - schema: {default: 1, format: int32, type: integer}
        in: query
        name: countryId
        required: false
      - schema: {default: '', type: string}
        in: query
        name: city
        required: false
      - schema: {default: 0, format: int32, type: integer}
        in: query
        name: offset
        required: false
      - schema: {default: 10, format: int32, type: integer}
        in: query
        name: itemPerPage
        required: false
      tags: [Контроллер для получения списка стран и подгрузки городов для страны]
...
  /api/v1/platform/countries:
    get:
      summary: Получение списка стран
      security:
      - jwt: []
      operationId: getCountries
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/ListResponsePlaceDto'}
      parameters:
      - schema: {default: '', type: string}
        in: query
        name: country
        required: false
      - schema: {default: 0, format: int32, type: integer}
        in: query
        name: offset
        required: false
      - schema: {default: 250, format: int32, type: integer}
        in: query
        name: itemPerPage
        required: false
      tags: [Контроллер для получения списка стран и подгрузки городов для страны]
  /api/v1/users/{id}:
    get:
      summary: Получение пользователя по его id
      security:
      - jwt: []
      operationId: getUserById
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/DataResponseAuthData'}
      parameters:
      - schema: {format: int32, type: integer}
        in: path
        name: id
        required: true
      tags: [Контроллер для работы с профилем]
...
  /api/v1/users/search:
    get:
      summary: Поиск пользователя
      security:
      - jwt: []
      operationId: search
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/ListResponseAuthData'}
      parameters:
      - schema: {default: '', type: string}
        in: query
        name: first_name
        required: false
      - schema: {default: '', type: string}
        in: query
        name: last_name
        required: false
      - schema: {default: -1, format: int32, type: integer}
        in: query
        name: age_from
        required: false
      - schema: {default: -1, format: int32, type: integer}
        in: query
        name: age_to
        required: false
      - schema: {default: '', type: string}
        in: query
        name: country
        required: false
      - schema: {default: '', type: string}
        in: query
        name: city
        required: false
      - schema: {default: 0, format: int32, type: integer}
        in: query
        name: offset
        required: false
      - schema: {default: 20, format: int32, type: integer}
        in: query
        name: itemPerPage
        required: false
      tags: [Контроллер для работы с профилем]
...
  /api/v1/tags:
    post:
      summary: Добавить тег
      security:
      - jwt: []
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TagDto'}
        required: true
      operationId: postTag
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/DataResponseTagDto'}
      tags: [Контроллер для работы с тегами]
    get:
      summary: Получить теги
      security:
      - jwt: []
      operationId: getTags
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/ListResponseTagDto'}
      parameters:
      - schema: {default: '', type: string}
        in: query
        name: tag
        required: false
      - schema: {default: 0, format: int32, type: integer}
        in: query
        name: offset
        required: false
      - schema: {default: 10, format: int32, type: integer}
        in: query
        name: itemPerPage
        required: false
      tags: [Контроллер для работы с тегами]
    delete:
      summary: Удалить тег
      security:
      - jwt: []
      operationId: deleteTag
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/AccountResponse'}
      parameters:
      - schema: {format: int32, type: integer}
        in: query
        name: id
        required: true
      tags: [Контроллер для работы с тегами]
...
  /api/v1/users/block/{id}:
    delete:
      summary: Разблокировать пользователя
      security:
      - jwt: []
      operationId: unBlockUser
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/AccountResponse'}
      parameters:
      - schema: {format: int32, type: integer}
        in: path
        name: id
        required: true
      tags: [Контроллер для работы с профилем]
    put:
      summary: Заблокировать пользователя
      security:
      - jwt: []
      operationId: blockUser
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema: {$ref: '#/components/schemas/AccountResponse'}
      parameters:
      - schema: {format: int32, type: integer}
        in: path
        name: id
        required: true
      tags: [Контроллер для работы с профилем]
...
info: {title: Zerone Social Network API, version: v1}
...